+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/Common/StartMeetingPresenter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using ICD.Common.Utils;
 using ICD.Common.Utils.EventArguments;
+using ICD.Common.Utils.Timers;
 using ICD.Connect.Calendaring;
 using ICD.Connect.Calendaring.Booking;
 using ICD.Connect.Calendaring.CalendarControl;
@@ -28,6 +29,7 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common
 
 		private readonly SafeCriticalSection m_RefreshSection;
 		private readonly ReferencedSchedulePresenterFactory m_ChildrenFactory;
+		private readonly SafeTimer m_RefreshTimer;
 
 		private IReferencedSchedulePresenter m_SelectedBooking;
 		private ICalendarControl m_CalendarControl;
@@ -46,8 +48,9 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common
 		{
 			m_RefreshSection = new SafeCriticalSection();
 			m_ChildrenFactory = new ReferencedSchedulePresenterFactory(nav, ItemFactory, Subscribe, Unsubscribe);
-
 			m_Bookings = new List<IBooking>();
+
+			m_RefreshTimer = SafeTimer.Stopped(UpdateBookings);
 		}
 
 		/// <summary>
@@ -55,6 +58,7 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common
 		/// </summary>
 		public override void Dispose()
 		{
+			m_RefreshTimer.Dispose();
 			m_ChildrenFactory.Dispose();
 
 			base.Dispose();
@@ -82,7 +86,6 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common
 				view.SetLogoPath(Theme.LogoAbsolutePath);
 
 				view.SetStartMyMeetingButtonEnabled(!HasCalendarControl || m_SelectedBooking != null);
-
 				view.SetStartNewMeetingButtonEnabled(HasCalendarControl);
 
 				if (HasCalendarControl && m_Bookings.Count < 1)
@@ -226,6 +229,8 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common
 
 		private void UpdateBookings()
 		{
+			IBooking first;
+
 			m_RefreshSection.Enter();
 
 			try
@@ -236,10 +241,23 @@ namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common
 						: m_CalendarControl.GetBookings()
 						                   .Where(b => b.EndTime > IcdEnvironment.GetLocalTime())
 						                   .ToList();
+
+				first = m_Bookings.FirstOrDefault();
 			}
 			finally
 			{
 				m_RefreshSection.Leave();
+			}
+
+			// Refresh when the next meeting starts or the current meeting ends.
+			if (first != null)
+			{
+				bool started = first.StartTime <= IcdEnvironment.GetLocalTime();
+				DateTime nextRefresh = started ? first.EndTime : first.StartTime;
+				long delta = (long)(nextRefresh - IcdEnvironment.GetLocalTime()).TotalMilliseconds + 1000;
+
+				if (delta > 0)
+					m_RefreshTimer.Reset(delta);
 			}
 
 			RefreshIfVisible();